% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_np_cols.R
\name{add_np_cols}
\alias{add_np_cols}
\title{Add `np` columns to a dataframe.}
\usage{
add_np_cols(df, np)
}
\arguments{
\item{df}{the dataframe to add columns to}

\item{np}{the number of columns to add.}
}
\value{
df
}
\description{
Representing the principal components a column for the PC score
  is added for each PC.
  The label for each column is given by the function \link{p_col_df}.
}
