% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_all_dist_pairs.R
\name{calc_all_dist_pairs}
\alias{calc_all_dist_pairs}
\title{Calculate the distance between each snp and every other snp.}
\usage{
calc_all_dist_pairs(score_mat, norm_typ = "F")
}
\arguments{
\item{score_mat}{the data matrix. Each row is a snp and the values
are the co-ordinates of the snp accross the trait axis.}

\item{norm_typ}{the type of norm to use in the distance calculations
when comparig the snps. Default is the Froebenius norm "F".}
}
\value{
dist_df
}
\description{
Using [calc_col_dist] and [calc_pair_dist]
  find the distance between each snp and every other snp.
}
\details{
"dist_df" is a dataframe with columns\:
  * "snp1" snp label
  * "snp2" snp label for the compared snp
  * "dist" the distance between snp1 and snp2
}
\seealso{
Other distance_functions: 
\code{\link{calc_clust_dist_col}()},
\code{\link{calc_clust_dist}()},
\code{\link{calc_col_dist}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{calc_pair_dist_df}()},
\code{\link{calc_snp_cent_dist}()},
\code{\link{max_dist_calc}()}
}
\concept{distance_functions}
