% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clust_ic.R
\name{calc_clust_ic}
\alias{calc_clust_ic}
\title{Find the BIC and AIC for the clusters.}
\usage{
calc_clust_ic(clust, group_col, dist_col, num_axis)
}
\arguments{
\item{clust}{The cluster membership dataframe.}

\item{group_col}{The name of the column to group terms on}

\item{dist_col}{The naame of the column to use distances from.}

\item{num_axis}{The number of axis i.e the dimensionality}
}
\value{
out
}
\description{

}
