% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_col_dist.R
\name{calc_col_dist}
\alias{calc_col_dist}
\title{Find the distance between a snp and all other snps.}
\usage{
calc_col_dist(score_mat, snp1, norm_typ = "F")
}
\arguments{
\item{score_mat}{Matrix of scores for each snp}

\item{snp1}{the snp name to compare all others to}

\item{norm_typ}{the type of norm to use in distance calculations.
default is the Froebenius norm "F".}
}
\value{
dist_df
}
\description{
Use [calc_pair_dist_df] to find the distance between pairs of snps.
  The results from this for each pair is bound on the rows using
  "rbind" into the datframe "dist_df".
}
