% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scale_mat.R
\name{calc_scale_mat}
\alias{calc_scale_mat}
\title{Rescale the columns of the matrix "mat"}
\usage{
calc_scale_mat(mat, narm = TRUE)
}
\arguments{
\item{mat}{The matrix whose columns are to be rescaled}

\item{narm}{Bool to indicate whether NaNs should be ignored in calculations.
default \: TRUE}
}
\value{
out_mat
}
\description{
Rescale the columns of the matrix `mat` so that the values lie
  between -1 and 1 with the with the same distribution.
}
\details{
Rescale of each column is found using [calc_col_scale] then
  column binding the results using cbind into "out_mat".
}
