% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_na.R
\name{check_col_na}
\alias{check_col_na}
\title{Check the percentage of NaNs in column}
\usage{
check_col_na(b_col, percent = 0.95)
}
\arguments{
\item{b_col}{- The column vector from the matrix}

\item{percent}{- The percentage of the column which must be not NaN}
}
\value{
0 or 1 - 1 if `b_col` has an acceptable number of non_NaN, else 0.
}
\description{
Check the percentage of NaNs in column
}
\seealso{
Other checking_functions: 
\code{\link{check_trait}()}
}
\concept{checking_functions}
