% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trait.R
\name{check_trait}
\alias{check_trait}
\title{Check is a trait is valid}
\usage{
check_trait(a, pheno_list, data_mat, na_percent)
}
\arguments{
\item{a}{- The label for the trait.}

\item{pheno_list}{- The list of all traits.}

\item{data_mat}{- The matrix containing the data of interest.
With columns labelled by trait labels.}

\item{na_percent}{- percentage of a column which is required to be NOT NaN.}
}
\value{
trait_single_df
}
\description{
if trait passes [na_col_check]:
    trait_single_df = data.frame(label = a,
                             axes_ind = position in pheno_list)
  else: trait_single_df = data.frame()
}
