% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clust_prob.R
\name{clust_prob_calc}
\alias{clust_prob_calc}
\title{Calculate the probability the snp is in the cluster}
\usage{
clust_prob_calc(d)
}
\arguments{
\item{d}{The snp distance to the cluster centre}
}
\value{
p
}
\description{
\deqn{p=\frac{1.0}{1.0+d}}
}
