% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_kmeans_basic.R
\name{cluster_kmeans_basic}
\alias{cluster_kmeans_basic}
\title{Cluster data with a standard kmeans approach}
\usage{
cluster_kmeans_basic(
  b_mat,
  nclust = 10,
  space_typ = "regular",
  clust_prob_on = TRUE,
  norm_typ = "F",
  threshold = 1e-05,
  narm = TRUE
)
}
\arguments{
\item{b_mat}{Matrix of data. Rows correspond to snps.}

\item{nclust}{Number of clusters to allocate}

\item{space_typ}{Whether to uses angles of spatial co-ordinates.
If "angle" use angles found using [convert_point_mat_to_angle]}

\item{clust_prob_on}{Bool switch. If TRUE then calculate the snps
probability of being in the cluster using [calc_clust_prob]}

\item{norm_typ}{The type of norm to use in distance calculations.
The default is the Froebenius norm "F".}

\item{threshold}{The threshold with which clusters centres must differ
by for clusters to be considered converged.}

\item{narm}{Boolean swithc. If TRUE (default) ignore NaNs in calculations.}
}
\value{
clusters_df
}
\description{
If space_typ == "angle" then data is converted to angles.
  The points are clustered using [km_nan] and "nclust" clusters.
  The "cluster_df" dataframe labelled
    "clusters" of the cluster membership with columns\:
    * "clust_num" the number of clusters
    * "clust_dist" the distance from the snp to the cluster centre
      (or distance between angles)
    * "clust_prob" probability the snp is in the cluster. Calculated
      using [calc_clust_prob].
}
\concept{clustering_functions}
