% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_cluster_list.R
\name{compare_oneclust_tolist}
\alias{compare_oneclust_tolist}
\title{Compare the cluster cn1 to all the clusters in the cluster list.}
\usage{
compare_oneclust_tolist(cn1, c_nums, c_score0, axis, clust_norm = "F")
}
\arguments{
\item{cn1}{- The number of the cluster being compared.}

\item{c_nums}{- The list of all the cluster numbers.}

\item{c_score0}{- The dataframe of cluster scores.}

\item{axis}{- The traits/ columns the clusters are being compared on.}

\item{clust_norm}{- The type of norm to be used when compareing the
difference accross multiple axis. Default is "F" which is the
Froebenius norm.}
}
\value{
diff_df
}
\description{
Output data frame of the difference in cluster
  scores between each pair.
}
\details{
As each difference is found the pair is added to the dataframe `diff_df`.
  This dataframe has columns "c_num1", "c_num2", "diff"
}
\concept{cluster_analysis_functions}
