% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_point_to_angle.R
\name{convert_point_to_angle}
\alias{convert_point_to_angle}
\title{Convert points to angle values}
\usage{
convert_point_to_angle(col, unit_mat, p)
}
\arguments{
\item{col}{The column on the unit matrix to find the angle to}

\item{unit_mat}{the matrix of unit vectors}

\item{p}{The point to find the angle for.}
}
\value{
\eqn{\theta}
}
\description{
Convert the point co-ordinates to values corresponding
  to angles to the unit vectors for each axis.
}
\details{
\deqn{
  \theta = \acos{\frac{p \cdot u}{||u|| ||p||}}.
  }
  If \eqn{\theta>\pi/2} then reassign to the accute angle of the extended
  vector.
  \deqn{\theta = \pi - \theta}
}
