% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_point_to_angles_all.R
\name{convert_point_to_angles_all}
\alias{convert_point_to_angles_all}
\title{Convert spatial co-ordinate to vector of angles}
\usage{
convert_point_to_angles_all(p, mat, unit_mat)
}
\arguments{
\item{p}{The point label to be converted}

\item{mat}{The matrix containing the co-ordinates for the points.}

\item{unit_mat}{The matrix of unit vectors in columns.}
}
\value{
ang_vec
}
\description{
Convert a point vector to a vector of values corresponding
  to the angles to the unit axis.
}
\details{
Calculate the angles using \link{convert_point_to_angle} for
  each column in unit_mat.
  Combine the angles into a vector "ang_vec"
}
