% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_full_desc_str.R
\name{create_full_desc_str}
\alias{create_full_desc_str}
\title{Create a sentence string describing the iteration.}
\usage{
create_full_desc_str(iter_traits)
}
\arguments{
\item{iter_traits}{- Data frame containing the iteration variables
"bp_on", "clust_prob_on",  "clust_typ"}
}
\value{
description_str
}
\description{
For titles and plot captions create a string with
 correct sentence structure which decribes the options in the iteration.
}
\details{
The string will start with the clustering type, then the bp switch,
 then the cluster probability switch.
}
