% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_col_names.R
\name{crop_col_names}
\alias{crop_col_names}
\title{Filter the dataframe and get the columns names that aren't in "ignore_cols"}
\usage{
crop_col_names(c_scores, filter_col, filter_val, ignore_cols)
}
\arguments{
\item{c_scores}{dataframe of interest}

\item{filter_col}{string label for column to filter by}

\item{filter_val}{value to filter "filter_col" with}

\item{ignore_cols}{list of column names to not include in output.}
}
\value{
col_list
}
\description{
* Filter the dataframe "c_scores" by the column "filter_col"
  with value "filter_val".
* Once filtered removed any all NaN columns.
* Extract the column names which aren't in "ignore_cols" to col_list
}
