% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_mat_colnums.R
\name{crop_mat_colnums}
\alias{crop_mat_colnums}
\title{Return the matrix with rows 1\:num_rows and columns col0\:col1}
\usage{
crop_mat_colnums(mat, num_rows, col0, col1)
}
\arguments{
\item{mat}{the matrix to be cropped}

\item{num_rows}{the number of rows to keep.}

\item{col0}{the first col to keep}

\item{col1}{the last col to keep}
}
\value{
cropped matrix
}
\description{
Return the matrix with rows 1\:num_rows and columns col0\:col1
}
