% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_ic.R
\name{find_all_ic}
\alias{find_all_ic}
\title{For each cluster find the aic and bic}
\usage{
find_all_ic(clust_re, num_axis)
}
\arguments{
\item{clust_re}{cluster results dataframe found using \link{km_nan}}

\item{num_axis}{the number of axis}
}
\value{
clust_re$clusters
}
\description{
Using \link{calc_clust_ic}
  for each cluster find the aic and bic and add this to the
  cluster results dataframe.
}
\details{
The aic and bic are found for each cluster using \link{calc_clust_ic}
  this is added to clust_re$clusters and clust_re$clusters is output.
  Ncents is also added as an axis to clust_re$clusters.
}
