% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eigen_vecs.R
\name{find_np_eigen_mat}
\alias{find_np_eigen_mat}
\title{Find the eigen values and vectors for the matrix mat.}
\usage{
find_np_eigen_mat(mat, np)
}
\arguments{
\item{mat}{the matrix to find the eigen values and vectors for.}

\item{np}{the number of eigen vectors to keep.}
}
\value{
e_mat
}
\description{
Take the np largest eigen values and their corresponding eigen vectors.
  Pairs with NaN correlation have no correlating scores and we can
  therefore set their correlation to 0.
}
\details{
"em_mat" is the matrix formed by column binding the eigen 
  vectors with "np" largest eigen values.
}
