% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_principal_components.R
\name{find_principal_components}
\alias{find_principal_components}
\title{Compute the np principal components of b_mat.}
\usage{
find_principal_components(b_mat, pval_mat, se_mat, np = 3, narm = TRUE)
}
\arguments{
\item{b_mat}{- Data matrix, columns are axis, rows are points.}

\item{pval_mat}{- P-values associated with each point.}

\item{se_mat}{- Standard errors of the data points.}

\item{np}{- Integer, number of prinipal components, default \:3}

\item{narm}{- Bool to indicate whether remove NaN values in
sample variables. default\: TRUE}
}
\value{
out_list
}
\description{
* Find the np largest Eigen values and eigen vectors.
  * Combine the vectors with column binding to form the transform
  matrix.
  * Transform the matrices into the princpal component space.
  * Combine the matrices into "out_list"
  \code{
  out_list   <- list("transform" = e_mat,
                 "beta" = b_pc_mat,
                  "pval" = pval_pc_mat,
                  "se" = se_pc_mat)
                  }
}
