% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_cluster_and_plot.R
\name{full_cluster_and_plot}
\alias{full_cluster_and_plot}
\title{Function to run clustering, scoring and plot the results.}
\usage{
full_cluster_and_plot(
  data_matrices,
  out_pheno,
  iter = 1,
  iter_df = data.frame(iter = 0, bp_on = FALSE, clust_prob_on = FALSE, clust_typ =
    "basic"),
  thresholds = list(threshmul = 5, diff = 1e-05, clust = 1e-05),
  na_handling = list(narm = TRUE, percent = 0.95),
  norm_typs = list(clust = "F", thresh = "F"),
  nums = list(max_dist = 1, np = 3, nr = 5)
)
}
\arguments{
\item{data_matrices}{- A list containing
* "beta" matrix corresponding to the data
* "pval" probabilities associated with "beta" values.
* "se" standard errors associated with "beta" values.
* "trait_info" dataframe of all the traits.}

\item{out_pheno}{- Label for the outcome phenotype}

\item{thresholds}{- List for the  threshold related variables.
* "threshmul" is multiplied by the data variance for cluster
difference threshold.
* "diff" is the the cluster difference threshold once found.
* "clust" is the required distance between cluster centres for a
cluster to be considered converged.}

\item{na_handling}{- List of the terms related to handling NaNs
* "narm" - TRUE (default) if NaNs are to be removed in calculations.
* "percent" - percentage of a column which must be not NaN. default\:0.95}

\item{norm_typs}{- List of norm types
* "clust" - default="F"
* "thresh" - default="F"}

\item{nums}{- List of important numbers.
* "max_dist" - Maximum distance between points, default=1
* "np" - Number of principal components
* "nc" - Number of clusters.}

\item{iter_traits}{- list of terms indicating the type of program.
* "iter" - default\:0
* "bp_on" - default\:TRUE switch if pval scores is to be used.
* "clust_prob_on" - default\:TRUE switch to use prob of being in cluster
* "clust_typ" - default="basic", the clustering method to use.
* "ndim_typ" - Whether to use all the traits or iterate through.}
}
\value{
df_list
}
\description{
Function to run clustering, scoring and plot the results.
}
\details{

}
