% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clust_col.R
\name{make_clust_col}
\alias{make_clust_col}
\title{Create a dataframe with column clust_i.}
\usage{
make_clust_col(i)
}
\arguments{
\item{i}{- The number for the column label}
}
\value{
dataframe(Pi = numeric())
}
\description{
Create a dataframe with column clust_i.
}
\details{
i is the input number
}
