% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_iter_df.R
\name{make_iter_df}
\alias{make_iter_df}
\title{Iteration setup function.}
\usage{
make_iter_df(clust_typ_list, bp_on_list, clust_prob_on_list, ndim_typ)
}
\arguments{
\item{clust_typ_list}{- List of the types of clustering to run on}

\item{bp_on_list}{- List of the variations of whether to include the
probability accoisated with the scores.}

\item{clust_prob_on_list}{- List of the variations of whether to
include the probability of whether a point is in a cluster based on
it's distance to the centre in cluster scoring.}

\item{ndim_typ}{- List of options for whether to iterate through
terms trait_list or run the program on them all.}
}
\value{
iter_df
}
\description{
Create a dataframe whose rows correspond to iterations of the
[full_cluster_and_plot] function. Each row indicates a different
set of input terms.
}
\details{
iter_df = data.frame() with columns: "bp_on", "clust_prob_on"
"ndim_typ", "clust_typ".
Each row is numbered and the dataframe contains all combinations
from the input lists.
}
