% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rand_cent.R
\name{make_rand_cent}
\alias{make_rand_cent}
\title{Create a set of randoms values}
\usage{
make_rand_cent(a, min_max_df, n_cents = 5)
}
\arguments{
\item{a}{the trait to create random values for.}

\item{min_max_df}{the dataframe of minimum and maximum
values for each trait. The rows are trait labelled and
the columns are "min" and "max"}

\item{n_cents}{the number of random values to create. default\:5}
}
\value{
cent_df
}
\description{
The points are selected using a uniform distribution
between the maximum and minimum values on the axis a.
}
\details{
Create a dataframe "cent_df" with column label "a" and the values are the
random values created using \link[stats]{runif}.
}
