% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unit_vec.R
\name{make_unit_vec}
\alias{make_unit_vec}
\title{Make a unit vector}
\usage{
make_unit_vec(i, tot_n)
}
\arguments{
\item{i}{the position for 1 to be located.}

\item{tot_n}{the total number of terms to be in the vector.}
}
\value{
u
}
\description{
Create vector of zeros "u",
allocate "i"th position of "u" to 1.
}
