% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mxdf.R
\name{plot_max_diff}
\alias{plot_max_diff}
\title{Plot the max difference against the number of traits.}
\usage{
plot_max_diff(max_df, iter_traits, pw = 16, ph = 4)
}
\arguments{
\item{max_df}{The dataframe of max difference scores and number of axis.}

\item{iter_traits}{the iteration options}

\item{pw}{the width of the plot. default\:16}

\item{ph}{the height of the plot. default \:4}
}
\description{
Plot the max difference against the number of traits.
}
