% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maxdf_list.R
\name{plot_mxdf_list}
\alias{plot_mxdf_list}
\title{Plot the maximum difference at each iteration for the
different input variables.}
\usage{
plot_mxdf_list(max_df, iter_traits, pw = 4, ph = 4)
}
\arguments{
\item{max_df}{dataframe of the maximum differences}

\item{iter_traits}{the iteration options for each iteration}

\item{pw}{the width of the plot. default\:4}

\item{ph}{the height of the plot. default\:4}
}
\description{
Plot the maximum difference at each iteration for the
different input variables.
}
