% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sngl_mxdf.R
\name{plot_sngl_mxdf}
\alias{plot_sngl_mxdf}
\title{Plot the maximum difference for a single iteration set.}
\usage{
plot_sngl_mxdf(plot_iter, max_df_list, iter_df, plotter)
}
\arguments{
\item{plot_iter}{the iteration options row.}

\item{max_df_list}{the list of max difference dataframes
accross all the iteration options}

\item{iter_df}{dataframe of iteration options}

\item{plotter}{Open plot to add to}
}
\value{
plot
}
\description{
Plot the maximum difference for a single iteration set.
}
