% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trait_heatmap.R
\name{plot_trait_heatmap}
\alias{plot_trait_heatmap}
\title{Plot a heatmap of the scores for each trait for each cluster.}
\usage{
plot_trait_heatmap(c_scores, iter_traits, pw = 16, ph = 4)
}
\arguments{
\item{c_scores}{- Dataframe of clusters and their scores for each axis.
Columns for dataframe are the traits and "clust_num". Rows each cluster
number and number of axis variation. Represented by nai_cnj, where i is
the number of axis and j is the cluster number.
More details in \link{score_trait}, \link{score_cluster}
and \link{score_all_clusters}}

\item{iter_traits}{- The properties of the iteration. List containing\:
* "bp_on",
* "clust_prob_on",
* "resdir",
* "clust_typ",
* "ndim_typ"
Destriptions for these terms in \link{make_trait_df}}

\item{pw}{Plot width, default\:16}

\item{ph}{Plot heigh, default\:4.}
}
\value{
Final heatmap plot
}
\description{
For values of num_axis plot the heatmap of traits against cluster score.
The limits on the axis are set on the range of the full dataset.
}
