% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transform_map.R
\name{plot_transform_heatmap}
\alias{plot_transform_heatmap}
\title{Plot the associations between traits and principal components.}
\usage{
plot_transform_heatmap(e_mat, iter_traits, num_axis = 0, pw = 16, ph = 4)
}
\arguments{
\item{e_mat}{the matrix of the PC transform}

\item{iter_traits}{the iteration details}

\item{num_axis}{the number of axis.}

\item{pw}{the width of the plot. default\:16}

\item{ph}{the height of the plot. default\:4}
}
\description{
Plot the associations between traits and principal components.
}
