% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cluster.R
\name{score_cluster}
\alias{score_cluster}
\title{Find the score for each trait for a given cluster.}
\usage{
score_cluster(c_num, beta_mat, clusters_df, num_axis, pval_mat, bp_on = TRUE)
}
\arguments{
\item{c_num}{- cluster number}

\item{beta_mat}{- data matrix}

\item{clusters_df}{- The cluster membership dataframe.
row names are the SNPs
Columns are:
* "clust_num" the number of the cluster the snp has been assigned to.
* "clust_prob" the probability the snp is in the cluster.
* "clust_dist" the distance to the cluster centre.}

\item{num_axis}{- The number of traits in use.}

\item{pval_mat}{- The matrix of probabilities for the data in beta_mat.}

\item{bp_on}{- Boolean, if TRUE the probabilities in pval_mat are used
in the cluster scoring. Default is TRUE}
}
\value{
cscore
}
\description{
Create a dataframe of the score for each trait using
\link{score_trait}. Column bind together into cscore
Add cluster number to column in cscore.
}
