% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_trait.R
\name{score_trait}
\alias{score_trait}
\title{Get the score for a cluster on the trait axis.}
\usage{
score_trait(beta_mat, pval_mat, snp_list, a, c_id, cprob_mat, bp_on = TRUE)
}
\arguments{
\item{beta_mat}{- Matrix of the data. Rows-snps, columns-traits.}

\item{pval_mat}{- Probabilities associated with beta_mat}

\item{snp_list}{- The list of snps in this cluster}

\item{a}{- The trait axis}

\item{c_id}{- The id for the cluster score.}

\item{cprob_mat}{- The probability the snp is in the cluster.}

\item{bp_on}{- Boolean switch, if TRUE (default)
use the pval_mat to weight the snp contribution to the cluster score.}
}
\value{
out
}
\description{
b_sub=Get the data at the snp_rows and trait column of beta_mat
c_sub=Get the data at the snp_rows and trait column of cprob_mat
Omit any NaNs.
if bp_on:
 p_sub = Get the data at the snp_rows and trait column of pval_mat
 out = b_sub*p_sub*c_sub/ (p_sub*c_sub)
else:
 out = b_sub * c_sub / (sum(c_sub))
}
