% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dir.R
\name{set_directory}
\alias{set_directory}
\title{Set the directory using the base directory and iteration vairables}
\usage{
set_directory(res_dir0, iter_traits)
}
\arguments{
\item{res_dir0}{- string for the base directory.}

\item{iter_traits}{- The dataframe containing the iteration variables.}
}
\value{
the directory path as a string
}
\description{
Create the string for the directory from the iteration variables using
the function \link{method_str}.
Combine the base directory and iteration directory.
Create the directory if it doesn't exist.
}
