% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_algorithm_data.R
\name{setup_algorithm_data}
\alias{setup_algorithm_data}
\title{Setup your data for running clustering functions}
\usage{
setup_algorithm_data(
  threshmul = 5,
  clust_threshold = 1e-05,
  na_percent = 0.8,
  nr = 10,
  np = 3,
  clust_norm = "F",
  thresh_norm = "F"
)
}
\arguments{
\item{threshmul}{The multiplier for the thresholding of the cluster
score difference. default\:5}

\item{clust_threshold}{The threshold for cluster centre difference to
determine whether clusters have converged.}

\item{na_percent}{The percentage of a column that must be not NaN.
default\:0.8}

\item{nr}{The maximum number of clusters. default\:10}

\item{np}{The number of principal components. default\:np}

\item{clust_norm}{The type of norm used for finding the
distance between points. default is the Froebenius norm "F".}

\item{thresh_norm}{The type of norm used for finding the
distance between scores. default is the Froebenius norm "F".}
}
\value{
empty
}
\description{
The function is written to use create variables that will be
needed globally for running the package functions.
}
\details{
Variables are created and declared globally.
thresholds <- ("threshmul" = threshmul,
                 "diff" = 1e-5,
                "clust" = clust_threshold)
na_handling <- ("narm" = TRUE, "percent" = na_percent)
nums <- ("nr" = nr, "np" = np, "max_dist" =  1.0)
norm_typs <- ("clust" =  clust_norm, "thresh_norm" =  thresh_norm)
}
