% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_na_check.R
\name{test_na_check}
\alias{test_na_check}
\title{Check the `na_col_check` function}
\usage{
test_na_check(b_mat, nan_col = "30600_irnt")
}
\arguments{
\item{b_mat}{- Data matrix with trait labelled columns}

\item{nan_col}{- Column label for an invalid column}
}
\value{
0,1 - 1 if work, 0 if not
}
\description{
Check the `na_col_check` function
}
