% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coords_in_mat.R
\name{transform_coords_in_mat}
\alias{transform_coords_in_mat}
\title{Transform the co-ordinates given by the rows in p_mat
to a coordinate system with basis given by the columns of t_mat}
\usage{
transform_coords_in_mat(p_mat, t_mat)
}
\arguments{
\item{p_mat}{the Matrix whos rows are points to be transformed}

\item{t_mat}{the transform matrix, rows correspond to the columns
of p_mat. Columns are the number of dimensioned beinng reduced to.}
}
\value{
vec_mat
}
\description{
For each row in p_mat multiply by t_mat. Bind the output vectors into
"vec_mat" using \link[SparkR]{rbind}
}
