% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_vec_by_mat.R
\name{transform_vec_by_mat}
\alias{transform_vec_by_mat}
\title{Transform a vector onto a new coordinate system using transform matrix}
\usage{
transform_vec_by_mat(p_mat, r, t_mat)
}
\arguments{
\item{p_mat}{Matrix whose rows are points}

\item{r}{The row of matrix corresponding to the point to be
transformed}

\item{t_mat}{The transform matrix whose rows correspond to the
columns of p_mat}
}
\value{
v
}
\description{
transform the vector p_mat[r,] to out_vec on a co-ordinate
  system whose basis is the columns of t_mat.
}
\details{
\deqn{v=p_mat[r,] * t_mat}
Any NaNs in "p_mat[r,]" are set to 0 before transform. Due to
dimensionality reduction it's not possible to maintain NaN tracking.
}
\examples{
transform_vec_by_mat(matrix(c(1,2,4,5,7,8), nrow=2, byrow=TRUE),
                     1,
                     matrix(c(1,2,4,5,7,8), nrow=3, byrow=TRUE))
}
