% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_cluster_and_plot.R
\name{full_cluster_and_plot}
\alias{full_cluster_and_plot}
\title{Function to run clustering, scoring and plot the results.}
\usage{
full_cluster_and_plot(
  data_matrices,
  out_pheno,
  res_dir0 = "",
  iter = 1,
  iter_df = data.frame(iter = 1, bp_on = FALSE, clust_prob_on = FALSE, clust_typ =
    "basic", ndim_typ = "all"),
  thresholds = list(threshmul = 5, diff = 1e-05, clust = 1e-05),
  na_handling = list(narm = TRUE, percent = 0.95),
  norm_typs = list(clust = "F", thresh = "F"),
  nums = list(max_dist = 1, np = 3, nr = 5)
)
}
\arguments{
\item{data_matrices}{- A list containing
* "beta" matrix corresponding to the data
* "pval" probabilities associated with "beta" values.
* "se" standard errors associated with "beta" values.
* "trait_info" dataframe of all the traits.}

\item{out_pheno}{- Label for the outcome phenotype}

\item{iter}{The iteratio number}

\item{iter_df}{- list of terms indicating the type of program.
* "iter" - default\:0
* "bp_on" - default\:TRUE switch if pval scores is to be used.
* "clust_prob_on" - default\:TRUE switch to use prob of being in cluster
* "clust_typ" - default="basic", the clustering method to use.
* "ndim_typ" - Whether to use all the traits or iterate through.}

\item{thresholds}{- List for the  threshold related variables.
* "threshmul" is multiplied by the data variance for cluster
difference threshold.
* "diff" is the the cluster difference threshold once found.
* "clust" is the required distance between cluster centres for a
cluster to be considered converged.}

\item{na_handling}{- List of the terms related to handling NaNs
* "narm" - TRUE (default) if NaNs are to be removed in calculations.
* "percent" - percentage of a column which must be not NaN. default\:0.95}

\item{norm_typs}{- List of norm types
* "clust" - default="F"
* "thresh" - default="F"}

\item{nums}{- List of important numbers.
* "max_dist" - Maximum distance between points, default=1
* "np" - Number of principal components
* "nc" - Number of clusters.}
}
\value{
df_list
}
\description{
Function to run clustering, scoring and plot the results.
}
\details{
based on iter_traits\:ndim_typ call the comparison and cluster function.
  If iter_trais\:ndim_typ=="all"{
    * run [clust_pca_all]
    * Plot the trait heatmap using [plot_trait_heatmap]
    * Plot the scatter plot using [plot_clust_scatter]
    * Plot the transform heatmap using [plot_transform_heatmap]
    } else{
    * run [clust_pca_compare_iterative]
    * Plot the heatmap for each iteration using [plot_trait_heatmap]
    * Plot the max difference through the iterations using
      [plot_max_diff]
    }
  df_list contains the results from all the iterations.
}
\seealso{
Other cluster_wrappers: 
\code{\link{clust_pca_all}()},
\code{\link{clust_pca_compare_iterative}()},
\code{\link{clust_pca_compare_single}()}
}
\concept{cluster_wrappers}
