% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_p_col.R
\name{make_p_col}
\alias{make_p_col}
\title{Create a dataframe with column Pi.}
\usage{
make_p_col(i)
}
\arguments{
\item{i}{- The number for the column label}
}
\value{
dataframe(Pi = numeric())
}
\description{
Create a dataframe with column Pi.
}
\details{
i is the input number
}
